Added speed setting in MenuShader module

+++ src/main/java/com/ferra13671/BThack/api/Utils/Ticker.java
@@ -7,7 +7,7 @@ public final class Ticker {
         time = -1;
     }
 
-    public float getPassedTime() {
+    public long getPassedTime() {
         return System.currentTimeMillis() - this.time;
     }
